requirements.txt
git+https://github.com/obstriker/WhatsappWebPy@main#egg=whatsapp_client
python-dotenv==1.0.0
agno==0.1.2
urllib3
requests
openai
sqlalchemy
#openai-whisper